refactor: exp points rtk slice and removed unnecessary code
